Move HydroShare related items to export
Move models, serializers, and the HydroShare service to the
`export` app: it doesn't belong in `modeling` because it is
not used for modeling.

Also update migrations to remake the table, and update any
references.
